parser_nistp/requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.1
cssselect==1.1.0
dnspython==1.16.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
mysql-connector-python==8.0.19
oauthlib==3.1.0
parsel==1.5.2
Pillow==7.1.1
Protego==0.1.16
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PySocks==1.7.1
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
Scrapy==2.0.1
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
tweepy==3.8.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
zope.interface==5.1.0
